Artificial_Intelligence/AI-Medical-Image-Analysis/requirements.txt
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
flask
streamlit
pillow
requests